move accordion to its own component
